separate tests of set_shots into standalone files
