+++ frontend/test/lit/test_device_api.py
@@ -20,7 +20,6 @@
 import os
 import pathlib
 import platform
-from functools import partial
 from typing import Optional
 
 import pennylane as qml
@@ -120,23 +119,3 @@ def test_preprocess():
 
 
 test_preprocess()
-
-
-def test_simple_circuit_set_shots():
-    """Test that a circuit with the new device API is compiling to MLIR."""
-    dev = qml.device("lightning.qubit", wires=2)
-
-    @qjit(target="mlir")
-    @partial(qml.set_shots, shots=2048)
-    @qml.qnode(device=dev)
-    def circuit():
-        qml.Hadamard(wires=0)
-        qml.CNOT(wires=[0, 1])
-        return qml.expval(qml.PauliZ(wires=0))
-
-    # CHECK: [[shots:%.+]] = arith.constant 2048 : i64
-    # CHECK: quantum.device shots([[shots]]) {{.*}}
-    print(circuit.mlir)
-
-
-test_simple_circuit_set_shots()
